Move Rule component to its own file
Also a minor fix in the multiple choice form.
